rename dto files to show models in swagger
